+++ fili-security/src/test/groovy/com/yahoo/bard/webservice/web/security/RoleBasedRoutingRequestMapperSpec.groovy
@@ -6,7 +6,6 @@ import com.yahoo.bard.webservice.data.config.ResourceDictionaries
 import com.yahoo.bard.webservice.web.ApiRequest
 import com.yahoo.bard.webservice.web.RequestMapper
 
-import spock.lang.Shared
 import spock.lang.Specification
 import spock.lang.Unroll
 
@@ -16,17 +15,20 @@ import javax.ws.rs.core.SecurityContext
 class RoleBasedRoutingRequestMapperSpec extends Specification {
 
     LinkedHashMap<String, RequestMapper<ApiRequest>> prioritizedRoleBasedMappers;
-    @Shared RequestMapper<ApiRequest> mapperA = Mock(RequestMapper)
-    @Shared RequestMapper<ApiRequest> mapperB = Mock(RequestMapper)
-    @Shared RequestMapper<ApiRequest> mapperC = Mock(RequestMapper)
+    RequestMapper<ApiRequest> mapperA = Mock(RequestMapper)
+    RequestMapper<ApiRequest> mapperB = Mock(RequestMapper)
+    RequestMapper<ApiRequest> mapperC = Mock(RequestMapper)
 
     ContainerRequestContext containerRequestContext = Mock(ContainerRequestContext)
     SecurityContext securityContext = Mock(SecurityContext)
     RoleBasedRoutingRequestMapper mapper
+    Map<String, RequestMapper<ApiRequest>> mappers = [a: mapperA, b: mapperB, c:mapperC]
 
     def setup() {
+        mapper = new RoleBasedRoutingRequestMapper(Mock(ResourceDictionaries), prioritizedRoleBasedMappers, mapperC)
         containerRequestContext.getSecurityContext() >> securityContext
         prioritizedRoleBasedMappers = [a: mapperA, b: mapperB]
+
         mapper = new RoleBasedRoutingRequestMapper(Mock(ResourceDictionaries), prioritizedRoleBasedMappers, mapperC)
     }
 
@@ -42,13 +44,21 @@ class RoleBasedRoutingRequestMapperSpec extends Specification {
         setup:
         ApiRequest apiRequest = Mock(ApiRequest)
         setupRoles(roles)
-        mapperB.apply(apiRequest, containerRequestContext) >> apiRequest
+        RequestMapper delegate = prioritizedRoleBasedMappers[nextMapper]
+
+        [mapperA, mapperB, mapperC].each {
+            int count = mappers[nextMapper] == it ? 1 : 0
+            count * it.apply(apiRequest, containerRequestContext) >> apiRequest
+        }
 
         expect:
         mapper.apply(apiRequest, containerRequestContext) == apiRequest
 
         where:
-        nextMapper | roles
-        mapperB    | ['b'] as Set
+        nextMapper   | roles
+        "a"          | ['a'] as Set
+        "b"          | ['b'] as Set
+        "a"          | ['a', 'b'] as Set
+        "c"          | [] as Set
     }
 }