+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Current
     * Add `availableIntervals` field to tables endpoint by union the availability for the logical table without taking
     the TablesApiRequest into account.
 
+- [Fili-security module](https://github.com/yahoo/fili/pull/405)
+    * Added security module for fili data security filters
+    * Extended Api 
+
 - [Implement EtagCacheRequestHandler](https://github.com/yahoo/fili/pull/312)
     * Add `EtagCacheRequestHandler` that checks the cache for a matching eTag
     * Add `EtagCacheRequestHandler` to `DruidWorkflow`